Renamed NamingConventionRule to ClassTypeNamingRule
