+++ extension.neon
@@ -7,5 +7,4 @@ parameters:
 		- stubs/Money/MoneyParser.stub
 rules:
     - Ibexa\PHPStan\Rules\NoConfigResolverParametersInConstructorRule
-    - Ibexa\PHPStan\Rules\RequireInterfaceInDependenciesRule
-    - Ibexa\PHPStan\Rules\NamingConventionRule
+    - Ibexa\PHPStan\Rules\ClassTypeNamingRule
