+++ tests/rules/Fixtures/ClassTypeNaming/SimpleClass.php
@@ -6,8 +6,8 @@
  */
 declare(strict_types=1);
 
-namespace Ibexa\Tests\PHPStan\Rules\Fixtures\NamingConvention;
+namespace Ibexa\Tests\PHPStan\Rules\Fixtures\ClassTypeNaming;
 
-interface CorrectNameInterface
+abstract class SimpleClass
 {
 }
