+++ tests/rules/Fixtures/ClassTypeNaming/WrongName.php
@@ -6,8 +6,8 @@
  */
 declare(strict_types=1);
 
-namespace Ibexa\Tests\PHPStan\Rules\Fixtures\NamingConvention;
+namespace Ibexa\Tests\PHPStan\Rules\Fixtures\ClassTypeNaming;
 
-abstract class AbstractCorrectClass
+interface WrongName
 {
 }
